+++ lua/catppuccin/groups/integrations/colorful_winsep.lua
@@ -1,0 +1,12 @@
+local M = {}
+
+function M.get()
+	return {
+		NvimSeparator = {
+			fg = C[O.integrations.colorful_winsep.color],
+			bg = O.transparent_background and C.none or C.base,
+		},
+	}
+end
+
+return M