+++ lua/catppuccin/groups/integrations/gitgutter.lua
@@ -5,6 +5,9 @@ function M.get()
 		GitGutterAdd = { fg = C.green },
 		GitGutterChange = { fg = C.yellow },
 		GitGutterDelete = { fg = C.red },
+		GitGutterAddLineNr = { fg = C.green },
+		GitGutterChangeLineNr = { fg = C.yellow },
+		GitGutterDeleteLineNr = { fg = C.red },
 	}
 end