+++ lua/catppuccin/groups/integrations/grug_far.lua
@@ -1,0 +1,9 @@
+local M = {}
+
+function M.get()
+	return {
+		GrugFarResultsMatch = { link = "IncSearch" },
+	}
+end
+
+return M