+++ lua/catppuccin/groups/integrations/nvim_surround.lua
@@ -1,0 +1,9 @@
+local M = {}
+
+function M.get()
+	return {
+		NvimSurroundHighlight = { sp = C.peach, style = { "underline" } },
+	}
+end
+
+return M