+++ lua/catppuccin/groups/integrations/treesitter.lua
@@ -17,13 +17,13 @@ If you want to stay on nvim 0.7, either disable the integration or pin catppucci
 		["@variable"] = { fg = C.text, style = O.styles.variables or {} }, -- Any variable name that does not have another highlight.
 		["@variable.builtin"] = { fg = C.red, style = O.styles.properties or {} }, -- Variable names that are defined by the languages, like this or self.
 		["@variable.parameter"] = { fg = C.maroon, style = O.styles.variables or {} }, -- For parameters of a function.
-		["@variable.member"] = { fg = C.lavender }, -- For fields.
+		["@variable.member"] = { fg = C.text }, -- For fields.
 
 		["@constant"] = { link = "Constant" }, -- For constants
 		["@constant.builtin"] = { fg = C.peach, style = O.styles.keywords or {} }, -- For constant that are built in the language: nil in Lua.
 		["@constant.macro"] = { link = "Macro" }, -- For constants that are defined by macros: NULL in C.
 
-		["@module"] = { fg = C.lavender, style = O.styles.miscs or { "italic" } }, -- For identifiers referring to modules and namespaces.
+		["@module"] = { fg = C.yellow, style = O.styles.miscs or { "italic" } }, -- For identifiers referring to modules and namespaces.
 		["@label"] = { link = "Label" }, -- For labels: label: in C and :label: in Lua.
 
 		-- Literals
@@ -34,7 +34,7 @@ If you want to stay on nvim 0.7, either disable the integration or pin catppucci
 		["@string.special"] = { link = "Special" }, -- other special strings (e.g. dates)
 		["@string.special.path"] = { link = "Special" }, -- filenames
 		["@string.special.symbol"] = { fg = C.flamingo }, -- symbols or atoms
-		["@string.special.url"] = { fg = C.rosewater, style = { "italic", "underline" } }, -- urls, links and emails
+		["@string.special.url"] = { fg = C.blue, style = { "italic", "underline" } }, -- urls, links and emails
 
 		["@character"] = { link = "Character" }, -- character literals
 		["@character.special"] = { link = "SpecialChar" }, -- special characters (e.g. wildcards)
@@ -45,11 +45,11 @@ If you want to stay on nvim 0.7, either disable the integration or pin catppucci
 
 		-- Types
 		["@type"] = { link = "Type" }, -- For types.
-		["@type.builtin"] = { fg = C.yellow, style = O.styles.properties or { "italic" } }, -- For builtin types.
+		["@type.builtin"] = { fg = C.mauve, style = O.styles.properties or { "italic" } }, -- For builtin types.
 		["@type.definition"] = { link = "Type" }, -- type definitions (e.g. `typedef` in C)
 
 		["@attribute"] = { link = "Constant" }, -- attribute annotations (e.g. Python decorators)
-		["@property"] = { fg = C.lavender, style = O.styles.properties or {} }, -- Same as TSField.
+		["@property"] = { fg = C.blue, style = O.styles.properties or {} }, -- Same as TSField.
 
 		-- Functions
 		["@function"] = { link = "Function" }, -- For function (calls and definitions).
@@ -60,7 +60,7 @@ If you want to stay on nvim 0.7, either disable the integration or pin catppucci
 		["@function.method"] = { link = "Function" }, -- For method definitions.
 		["@function.method.call"] = { link = "Function" }, -- For method calls.
 
-		["@constructor"] = { fg = C.sapphire }, -- For constructor calls and definitions: = { } in Lua, and Java constructors.
+		["@constructor"] = { fg = C.yellow }, -- For constructor calls and definitions: = { } in Lua, and Java constructors.
 		["@operator"] = { link = "Operator" }, -- For any operator: +, but also -> and * in C.
 
 		-- Keywords
@@ -69,7 +69,7 @@ If you want to stay on nvim 0.7, either disable the integration or pin catppucci
 		["@keyword.type"] = { link = "Keyword" }, -- For keywords describing composite types (e.g. `struct`, `enum`)
 		["@keyword.coroutine"] = { link = "Keyword" }, -- For keywords related to coroutines (e.g. `go` in Go, `async/await` in Python)
 		["@keyword.function"] = { fg = C.mauve, style = O.styles.keywords or {} }, -- For keywords used to define a function.
-		["@keyword.operator"] = { link = "Operator" }, -- For new keyword operator
+		["@keyword.operator"] = { fg = C.mauve, style = O.styles.keywords or {} }, -- For new keyword operator
 		["@keyword.import"] = { link = "Include" }, -- For includes: #include in C, use or extern crate in Rust, or require in Lua.
 		["@keyword.repeat"] = { link = "Repeat" }, -- For keywords related to loops.
 		["@keyword.return"] = { fg = C.mauve, style = O.styles.keywords or {} },
@@ -81,8 +81,9 @@ If you want to stay on nvim 0.7, either disable the integration or pin catppucci
 
 		["@keyword.directive"] = { link = "PreProc" }, -- various preprocessor directives & shebangs
 		["@keyword.directive.define"] = { link = "Define" }, -- preprocessor definition directives
+		["@keyword.directive.css"] = { link = "Keyword" }, -- CSS at-rules: https://developer.mozilla.org/en-US/docs/Web/CSS/At-rule.
 		-- JS & derivative
-		["@keyword.export"] = { fg = C.sky, style = O.styles.keywords },
+		["@keyword.export"] = { fg = C.mauve, style = O.styles.keywords },
 
 		-- Punctuation
 		["@punctuation.delimiter"] = { link = "Delimiter" }, -- For delimiters (e.g. `;` / `.` / `,`).
@@ -101,25 +102,26 @@ If you want to stay on nvim 0.7, either disable the integration or pin catppucci
 
 		-- Markup
 		["@markup"] = { fg = C.text }, -- For strings considerated text in a markup language.
-		["@markup.strong"] = { fg = C.maroon, style = { "bold" } }, -- bold
-		["@markup.italic"] = { fg = C.maroon, style = { "italic" } }, -- italic
+		["@markup.strong"] = { fg = C.red, style = { "bold" } }, -- bold
+		["@markup.italic"] = { fg = C.red, style = { "italic" } }, -- italic
 		["@markup.strikethrough"] = { fg = C.text, style = { "strikethrough" } }, -- strikethrough text
 		["@markup.underline"] = { link = "Underlined" }, -- underlined text
 
-		["@markup.heading"] = { fg = C.blue, style = { "bold" } }, -- titles like: # Example
+		["@markup.heading"] = { fg = C.blue }, -- titles like: # Example
+		["@markup.heading.markdown"] = { style = { "bold" } }, -- bold headings in markdown, but not in HTML or other markup
 
 		["@markup.math"] = { fg = C.blue }, -- math environments (e.g. `$ ... $` in LaTeX)
-		["@markup.quote"] = { fg = C.maroon, style = { "bold" } }, -- block quotes
+		["@markup.quote"] = { fg = C.pink }, -- block quotes
 		["@markup.environment"] = { fg = C.pink }, -- text environments of markup languages
 		["@markup.environment.name"] = { fg = C.blue }, -- text indicating the type of an environment
 
-		["@markup.link"] = { link = "Tag" }, -- text references, footnotes, citations, etc.
-		["@markup.link.label"] = { link = "Label" }, -- link, reference descriptions
-		["@markup.link.url"] = { fg = C.rosewater, style = { "italic", "underline" } }, -- urls, links and emails
+		["@markup.link"] = { fg = C.lavender }, -- text references, footnotes, citations, etc.
+		["@markup.link.label"] = { fg = C.lavender }, -- link, reference descriptions
+		["@markup.link.url"] = { fg = C.blue, style = { "italic", "underline" } }, -- urls, links and emails
 
-		["@markup.raw"] = { fg = C.teal }, -- used for inline code in markdown and for doc in python (""")
+		["@markup.raw"] = { fg = C.green }, -- used for inline code in markdown and for doc in python (""")
 
-		["@markup.list"] = { link = "Special" },
+		["@markup.list"] = { fg = C.teal },
 		["@markup.list.checked"] = { fg = C.green }, -- todo notes
 		["@markup.list.unchecked"] = { fg = C.overlay1 }, -- todo notes
 
@@ -129,9 +131,10 @@ If you want to stay on nvim 0.7, either disable the integration or pin catppucci
 		["@diff.delta"] = { link = "diffChanged" }, -- deleted text (for diff files)
 
 		-- Tags
-		["@tag"] = { fg = C.mauve }, -- Tags like html tag names.
-		["@tag.attribute"] = { fg = C.teal, style = O.styles.miscs or { "italic" } }, -- Tags like html tag names.
-		["@tag.delimiter"] = { fg = C.sky }, -- Tag delimiter like < > /
+		["@tag"] = { fg = C.blue }, -- Tags like HTML tag names.
+		["@tag.builtin"] = { fg = C.blue }, -- JSX tag names.
+		["@tag.attribute"] = { fg = C.yellow, style = O.styles.miscs or { "italic" } }, -- XML/HTML attributes (foo in foo="bar").
+		["@tag.delimiter"] = { fg = C.teal }, -- Tag delimiter like < > /
 
 		-- Misc
 		["@error"] = { link = "Error" },
@@ -152,12 +155,12 @@ If you want to stay on nvim 0.7, either disable the integration or pin catppucci
 		["@constant.java"] = { fg = C.teal },
 
 		-- css
-		["@property.css"] = { fg = C.lavender },
-		["@property.id.css"] = { fg = C.blue },
+		["@property.css"] = { fg = C.blue },
+		["@property.id.css"] = { fg = C.yellow },
 		["@property.class.css"] = { fg = C.yellow },
 		["@type.css"] = { fg = C.lavender },
-		["@type.tag.css"] = { fg = C.mauve },
-		["@string.plain.css"] = { fg = C.peach },
+		["@type.tag.css"] = { fg = C.blue },
+		["@string.plain.css"] = { fg = C.text },
 		["@number.css"] = { fg = C.peach },
 
 		-- toml
@@ -167,18 +170,11 @@ If you want to stay on nvim 0.7, either disable the integration or pin catppucci
 		["@label.json"] = { fg = C.blue }, -- For labels: label: in C and :label: in Lua.
 
 		-- lua
-		["@constructor.lua"] = { fg = C.flamingo }, -- For constructor calls and definitions: = { } in Lua.
-
-		-- typescript
-		["@property.typescript"] = { fg = C.lavender, style = O.styles.properties or {} },
-		["@constructor.typescript"] = { fg = C.lavender },
-
-		-- TSX (Typescript React)
-		["@constructor.tsx"] = { fg = C.lavender },
-		["@tag.attribute.tsx"] = { fg = C.teal, style = O.styles.miscs or { "italic" } },
+		["@constructor.lua"] = { link = "@punctuation.bracket" }, -- For constructor calls and definitions: = { } in Lua.
 
 		-- yaml
 		["@variable.member.yaml"] = { fg = C.blue }, -- For fields.
+		["@variable.member.nix"] = { fg = C.blue },
 
 		-- Ruby
 		["@string.special.symbol.ruby"] = { fg = C.flamingo },
@@ -188,12 +184,13 @@ If you want to stay on nvim 0.7, either disable the integration or pin catppucci
 		["@function.method.call.php"] = { link = "Function" },
 
 		-- C/CPP
-		["@type.builtin.c"] = { fg = C.yellow, style = {} },
 		["@property.cpp"] = { fg = C.text },
-		["@type.builtin.cpp"] = { fg = C.yellow, style = {} },
 
 		-- gitcommit
 		["@comment.warning.gitcommit"] = { fg = C.yellow },
+
+		-- gitignore
+		["@string.special.path.gitignore"] = { fg = C.text },
 
 		-- Misc
 		gitcommitSummary = { fg = C.rosewater, style = O.styles.miscs or { "italic" } },